Split main into handler and tokenizer
