+++ go-autocomplete/go.mod
@@ -2,4 +2,7 @@ module go-autocomplete
 
 go 1.24.3
 
-require github.com/razzie/onnxruntime-go v0.10.0
+require (
+    github.com/labstack/echo/v4 v4.11.4
+    github.com/razzie/onnxruntime-go v0.10.0
+)
